add pair attack
